Worked on dynamic routing
